Implement NewWriter
