+++ error.go
@@ -5,6 +5,7 @@ import (
 )
 
 var (
+	ErrMultipleWriteNotSupported = errors.New("multiple write not supported")
 	ErrMultiparterNotImplemented = errors.New("multiparter not implemented")
 	ErrAppenderNotImplemented    = errors.New("appender not implemented")
 )
